requirements.txt
black==24.8.0
django-jazzmin==3.0.0
isort==5.13.2
pillow==10.4.0
pip-chill==1.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
